Expose print identifier helper
